Refactor: move set-repair-config definitions to command::set_repair_config
