+++ src/lib.rs
@@ -35,6 +35,8 @@ extern crate slog;
 extern crate tempdir;
 #[macro_use]
 extern crate trackable;
+extern crate clap;
+extern crate sloggers;
 
 use std::fs::File;
 use std::net::SocketAddr;
@@ -49,6 +51,7 @@ macro_rules! dump {
 
 pub use error::{Error, ErrorKind};
 
+pub mod command;
 pub mod daemon;
 
 /// The following module is automatically generated by build.rs .
